Open graph meta tags are wrong and lead to problems
The open graph meta tags in config.tsx should implement the 'property' attribute instead of 'name' to be fully compliant with open graph requirements. I noticed that LinkedIn image preview is not working when using the 'name' attribute.
Also Facebook is complaining and stating
"The 'og:image' property should be explicitly provided, even if a value can be inferred from other tags."

These issues go away after switching to 'property' on the OG meta tags. Twitter, however, is fine with using the 'name' attribute which is confirmed by their documentation.